refactor: use event delegation for attack listener
Move attack listener one level up, to the parent container. This
ensures the attack listeners are not lost upon re-render. This
change was needed to implement the computer's attacks. Event
delegation was being used previously as well, but it was partial
delegation. We now have persistent delegation with this refactor.